Add Runtime name to Custom Breakpad Data
Summary:
Another thing I found difficult in my last investigation of fatals was determining whether the heap belonged to RN or CS.  This diffs logs the runtime's name when we OOM to help us differentiate when looking at the trace.

In the spirit of reducing memory usage, I also clamped the size of the name, to prevent us overrunning our buffer.  This may be somewhat in vain, however, because I noticed that `setBreakpadCustomData` is copying strings all over the place (for example, every time a key-value pair is added, it materialises all the custom data accumulated so far into a JSON string, to update Breakpad with). cc ddetlefs

Reviewed By: dulinriley

Differential Revision: D13659470

fbshipit-source-id: 90ca5b88b4c25eb15a10fbfd531804c3c38b3cea