LogFailStorageProvider
Summary:
Create an adapter `StorageProvider` that keeps track of how many times we have failed to allocate storage for a segment.

In a diff to follow, I will use this provider to count materialisation failures in NCGen.

There already existed an adapter that counted successful allocations for testing purposes -- `CountingStorageProvider`.  This diff renames that to `LogSuccessStorageProvider` to highlight the difference between it and the newly introduced adapter.

Reviewed By: dulinriley

Differential Revision: D13704744

fbshipit-source-id: 857cca5339a3c227d220caea954e6d3aaa52c644