Keep track of segment materialisation failures.
Summary:
Use LoggingStorageProvider in NCGen to track materialisation failures.

Also created a new test case for the scenario where we fail to materialise a segment but are able to continue anyway (because a full collection freed up enough space).

In a diff to follow, I will add this information to the custom breakpad data.

Reviewed By: dulinriley

Differential Revision: D13704749

fbshipit-source-id: 5f3defc6efc7e3ebf9cd88f49d047e026c4abb0b