Emit and implement Call1 through Call4 instructions
Summary:
This adds four new bytecode instructions, Call1 through Call4, all of
which are emitted by the CallN IR instruction. These bytecode instructions
work by copying their register arguments to the end of the frame, and then
invoking Call - effectively doing the Movs that a normal Call would
expect. Because the optimization pass is not enabled by default, these
instructions are not yet emitted (without specifying a custom pass, which
is what the tests do).

Because these instruction are marked as ret targets in the
BytecodeList.def, Ret does not need to be updated.

Reviewed By: tmikov

Differential Revision: D13561291

fbshipit-source-id: 6395474ac124ea641bdc1fedf85734a3d8c2429a