Add HBCCallN IR instructions and pass
Summary:
This adds a new IR instruction HBCCallN, which represents a Call where the
argument count is known and small (currently 1-4). The arguments are all
registers. The bytecode implementation will copy the registers to the end
of the frame (performing the same sort of logic as Call lowering) and then
do the usual Call thing. Because the parameter registers are overwritten,
as part of lowering, we emit ImplicitMovs instead of Movs to reflect the
copying that HBCCallN will do.

HBCCallN does not require its arguments to be registers at the end of the
frame; in fact it prohibits exactly that because it overwritse those
registers. Update verifyCall to reflect that.

A new optimization pass FuncCallNOpts replaces Call instructions with
HBCCallN when in range. ISel is not yet implemented for this new
instruction, and the pass is not yet enabled.

This is a second attempt at D13561290 which was backed out due to a
failing CMake test.

Reviewed By: tmikov

Differential Revision: D13729757

fbshipit-source-id: 42a6bb23b22d61a0a21237ed53ff252eb47ade39