Add ImplicitMov IR instruction
Summary:
This adds a new "ImplicitMov" IR instruction that represents a Mov
performed implicitly by a subsequent instruction.

ImplicitMov will be used by the CallN family of instructions. CallN copies
its operands to the argument passing registers, effectively as if by a
series of Mov instructions. If the Optimizer were ignorant of this
property, it would assume that the argument registers are unmodified by
CallN, and perform invalid Mov-elimination or other bogus optimizations
using these registers.

Example code (here f1 is a normal call and f2 is a HBCCallN):

  f1(1)
  f2(77)
  f1(1)

Potential post-ra (assuming we have 16 regs):

  %1 = HBCLoadConstInst 1
  %2 = MovInst %1 (to reg 15)
  %3 = Call %2
  %4 = HBCLoadConstInst 77
  %5 = HBCCallN %4
  %6 = MovInst %1 (to reg 15)
  %7 = Call %6

Here the optimizer might conclude that register 15 (holding the constant
1) is unchanged by the HBCCallN, and so conclude that the Mov %6 is
unnecessary. But actually f2 implicitly performed a Mov into register
15, so %6 is in fact necessary.

With an ImplicitMov we will produce the following:

  %1 = HBCLoadConstInst 1
  %2 = MovInst %1 (to reg 15)
  %3 = Call %2
  %4 = HBCLoadConstInst 77
  %5 = ImplicitMovInst %4 (to reg 15)
  %6 = HBCCallN %4
  %7 = MovInst %1 (to reg 15)
  %8 = Call %6

and now the optimizer will correctly preserve the last MovInst.

Note ImplicitMov does not emit any bytecode; it acts only as an optimization
barrier..

This is a second attempt at D13561289, which was backed out due to a CMake
test failure.

Reviewed By: tmikov

Differential Revision: D13729758

fbshipit-source-id: 5d97d9cc17687143ae4f61ba85ba4ef25ab0f6d4