Abstract "Ret-target" bytecode instructions
Summary:
In preparation for introducing CallN bytecode instructions, add some
abstraction around the Call family of instructions (Call, Construct,
CallLong, etc.).

Add a new macro DEFINE_RET_TARGET in BytecodeList.def, which calls out the
instructions which may be the savedIP inspected by a Ret. Statically
assert that these targets have monotone-increasing instruction sizes. This
replaces a manually maintained pair of assertions.

This is a second attempt at D13561287 which was backed out due to
a CMake test failure. The test failure occurred in this diff and was fixed
by declaring a LLVM_ATTRIBUTE_ALWAYS_INLINE function as also inline.

Reviewed By: tmikov

Differential Revision: D13729759

fbshipit-source-id: b3079c935b75e1e424fe179a24f4ec2e34c3f064