Emit and implement Call1 through Call4 instructions
Summary:
This adds four new bytecode instructions, Call1 through Call4, all of
which are emitted by the CallN IR instruction. These bytecode instructions
work by copying their register arguments to the end of the frame, and then
invoking Call - effectively doing the Movs that a normal Call would
expect. Because the optimization pass is not enabled by default, these
instructions are not yet emitted (without specifying a custom pass, which
is what the tests do).

Because these instruction are marked as ret targets in the
BytecodeList.def, Ret does not need to be updated.

This is a second attempt at D13561291 which was backed out due to a CMake
test failure.

Reviewed By: tmikov

Differential Revision: D13729756

fbshipit-source-id: 5f0caa630a0c8a7b18e00011fa4a35c605b399e9