Use C99 compliant syntax for flexible array member
Summary:
This commit changes the declaration of flexible array member in
JSONHiddenClass to be compliant with C99.

In addition, there is a practical reason. With  `JSONString *keys_[1]`,
in debug build, MSVC will add bound checks when executing
`std::copy(b, e, keys_)`, which fails when there are more than one
elements in the source iterator.

`void *operator new` is static regardless of whether it is
explicitly declared as such. The `static` specifier is added here
to make the code more obvious.

Reviewed By: tmikov

Differential Revision: D13755555

fbshipit-source-id: 90115bac0097c760df5c18970d5465f81539e9ab