Update xplat/hermes LLVM references to xplat/hermes/facebook
Summary:
We decided to move our copy of LLVM under the Hermes tree, to emphasize that it's an abridged version that we'll be updating only if we need to for some reason, not a general-purpose copy of llvm that other packages can depend on.

The previous diff in this stack does that move (actually, so far, a copy).  This diff updates all the references within hermes and fbandroid tree to this new location.

This does *not* update the occurrences of "third-party/llvm" in cmake_install.cmake scripts.  My understanding (possibly wrong!) is that these were used once to create the hg version from a git repo, as per (new file) xplat/hermes/facebook/llvm/README.facebook.  If this is wrong, and there's something I should be doing with these, please let me know!

[Apologies about the number of versions necessary -- please ignore everything before V7.  I had the copy and the modifications together, wanted to split them for easier reviewing, ran into some hg/arc confusions, eventually resolved it.)

allow-large-files

Reviewed By: mhorowitz

Differential Revision: D13600791

fbshipit-source-id: 2722dcea2c0941e70817782a263aa286ef64d08e