Enable grouping of AST nodes via inheritance
Summary:
Until now all AST nodes were flat and there was no way to group them
other than by explicitly enumerating all cases, which was very clumsy,
inefficient, not to mention error prone.

Enable AST nodes to inherit from abstract base nodes defined in the same
ESTree.def file. This enables us to use the familiar LLVM mechanics to
do this like this for example:

    if (isa<ESTree::LoopStatementNode>(node))
        ....;

Previously we had to explicitly list all loop nodes.

Additionally now we can decorate the base abstract node instead of
defining a decoration for every concrete node type. This turns out to be
a big deal when we want to access the same decoration in a "class" of
nodes. Before we would have to do something like this:

    LoopDecorationBase * decoration = nullptr;
    if (auto *DW = dyn_cast<DoWhileStatementNode>(node))
        decoration = DW;
    else if (auto *W = dyn_cast<WhileStatementNode>(node))
        decoration = W;
    else if (auto *F = dyn_cast<ForStatementNode>(node))
        decoration = F;
    else if ... // ... and so on

Now we can simply do:

    LoopStatementDecoration * decoration = nullptr;
    if (auto *LS = dyn_cast<LoopStatementNode>(node))
        decoration = LS;
    // DONE!

Reviewed By: avp

Differential Revision: D13595851

fbshipit-source-id: 617e189bc570aa747acf4a1657c40fbfee1606bf