Clean up ESTreeIRGen::genClosure()
Summary:
There are no functionality changes in this diff. It is a small
incremental cleanup. We used to have two functions called
`genClosure()`, where one was a template. It was very confusing,
especially since they didn't actually produce a closure.

Rename the higher level one to `genFunctionLike()`. It follows the
pattern of the other generation functions like `genExpression()`, etc.

Rename the lower level one to `doGenFunctionLike()`. That name is not
ideal, but I couldn't come up with a better one. At least it is now
different.

Also remove the template parameter and use `std::function<>`
instead. `doGenFunctionLike()` is not small, so we don't want more than
one copy of it.

Reviewed By: avp

Differential Revision: D13595841

fbshipit-source-id: 83ac58b63c053fddcbc737acea4f1de353795feb