Associate semantic information with function AST nodes
Summary:
This sets up the foundation for moving the construction of various IRGen
tables from IRGen to SemanticValidator. The validator needs to construct
a version of the same information anyway, in order to validate the AST.

Currently IRGen maintains several stacks and contextual tables for
labels, try/catch, variables, etc (this is likely to become even more
complex as we add more ES6 support). The end goal is to instead
eliminate all that from IRGen and make it simple as possible, relying on
information encoded in the AST during validation.

To that end we introduce `SemanticContext`, which is a repository for
all semantic information that will be associated with the AST. Why don't
we simply encode the information directly in the AST? For performance
the AST is allocated in a pool and nodes cannot have destructors, so
anything with a destructor has to be allocated separately.

The `SemanticContext` turns out to be very convenient for things like
lazy parsing, since we can simply construct it locally in the stack and
discard it when no longer needed.

Allocate a data structure sem::FunctionInfo and associate it with
Function AST nodes. It is empty for now, but will contain more and more
information in future diffs.

Lastly, move the semantic validator into a new namespace `hermes::sem`.

Reviewed By: willholen, avp

Differential Revision: D13617581

fbshipit-source-id: 3f1a4e74b239a2559cc2ac40304396c4ebc11181