simplify handling of try/catch/finally
Summary:
`try` statements with both `catch` and `finally` handlers are complex to
generate code for, since they technically are two nested `try/catch`
statements: an exception thrown in the `catch` handler must still invoke
the `finally` handler.

Let's use this input as an example:

    try {
        tryBody;
    } catch (e) {
        catchBody;
    } finally {
        finallyBody;
    }

Until now we were transforming it into the following pseudocode (it was
done implicitly by IRGen while emitting IR):

    try {
        tryBody;
    } catch (e) {
        try {
            catchBody;
        } catch (e) {
            finallyBody;
            throw e;
        }
    }
    finallyBody;

Now instead we are performing a simpler classical transformation:

    try {
        try {
            tryBody;
        } catch (e) {
            catchBody;
        }
    } finally {
        finallyBody;
    }

We can do that purely by AST manipulation during validation, leaving
IRGen to deal only with the simple case.

This leads to another welcome simplification of IRGen, making it a little
easier to reason about. That made it possible to tweak the debug location
information and make it more intuitive. Before it would jump back and
forth between the end and beginning of the `try` statement, while now it
just flows naturally.

Finally, this produces fewer instructions since there are fewer copies
of the finally handlers.

Reviewed By: avp

Differential Revision: D13595849

fbshipit-source-id: 4efd7200a49ce8e0ba6efb3a32e88b6b4a9b81df