resolve all gotos during validation
Summary:
Move the resolution of break/continue targets from IRGen to the semantic
validator, which needs to track them anyway in order to report errors.
Note that label resolution necessarily includes information about nested
`try/catch` statements, since the finally handlers need to be executed
when leaving a block.

This leads to a nice simplification of IRGen and, more importantly,
fixes a critical compiler bug - "Inlining of finally handlers uses
incorrect context". IRGen no longer needs to maintain break/continue
context, so it is always safe to inline the "finally" handler (at least
as far as labels are concerned).

What information is encoded? Each "goto target" is represented as an
integer number. The corresponding `break/continue` AST nodes are
decorated with that number, and so is the AST node defining the target
itself (things like loops, labeled statements). `try/catch` statements
are decorated with a pointer to the enclosing `try/catch`, and so are
`break/continue/return` statements and the "goto targets" themselves.
This allows a trivial algorithm to determine which `finally` handlers
need to be executed when performing a branch. See
`ESTreeIRGen::genFinallyBeforeControlChange()`.

Reviewed By: avp

Differential Revision: D13595843

fbshipit-source-id: 1b51a80bba85beaaf03388df79f6efa30e8ed62d