Introducing (for|while)UsedSegments
Summary:
Introducing a replacement to the `usedSegments()` API -- a family of iteration functions over segments that accept a callback parameter.  This is so that `OldGenSegmentIterator`, `SegTraits` etc can be removed when its usage in `allSegments()` is replaced by an interface that allows us to recover from a failure to materialise an old gen segment during full collection.

The interface presented in this diff was arrived upon through experimentation in godbolt: https://godbolt.org/z/SA6M5j, to ensure an interface that would compile down to assembly roughly comparable to manualy iterating through the segments.  I also tried an interface accepting a `const std::function<void(AlignedHeapSegment &)>` callback parameter which seemed to introduce considerable overhead.

Reviewed By: dulinriley

Differential Revision: D13755283

fbshipit-source-id: ae720d95b3f007cc1f9bcd7514eda2ea0b522243