PerfSection log event start and stop if platform logging enabled.
Summary: We used to only log the ends of events, if they had had arguments attached.  I was bitten by this recently, because I was expecting to see log messages for new invocations I had added, but they were not appearing because those invocations did not have argumetns attached.  This diff adds logs to the start and end of every event.

Reviewed By: dulinriley

Differential Revision: D13806444

fbshipit-source-id: e537521a32e96138b22d342d22e65bd48e65dfac