Back out "Update xplat/hermes LLVM references to xplat/hermes/facebook"
Summary:
Original commit changeset: 2722dcea2c09

(Diff created by hg backout 2722dcea2c0941e70817782a263aa286ef64d08e)

The original diff moved Hermes to depend on its internal copy of LLVM.  Between the time I originally created that diff, and the time I landed, some new references to xplat/third-party/LLVM crept in, in the Android merge maps.

Doing this backout, so we're sure we fix the problem, then I'll redo the diff with the new references also changed, and make sure that the build bot report is neutral before publishing.

ADDENDUM: it turns out we need to *not* back out a couple of the changes.  I want to leave the previous diff that just did the copy of the source to the new location, but changed no BUCK files to reference them (that's what the backed-out diff does) in place.  But some tests, the "cmake check" tests, traverse the source tree, making sure that all the files are mentioned in CMakeList.txt files.  We didn't do these for LLVM directories before; we don't want do those directories now, even though they're under the xplat/hermes directory.  So that's why the additional changes in the cmake_check.sh files (to the versions in the backed-out diff, which exempts the LLVM files.)

Reviewed By: zlern2k

Differential Revision: D13805694

fbshipit-source-id: deaf835d4ec63b1e6be768ecd56623bba19253f1