keep track of "new.target"
Summary:
`new.target` is an ES6 meta-property that contains the callable of the
constructor which is currently being invoked by `new`, or `undefined` if
it is a normal function call.

It supersedes the previous `construct` parameter which we used to pass
around in the VM.

It is implemented as a reserved field in the stack frame. That is not
ideal, because we are increasing the size of all function stack frames
by one, while at the same time few (if any at all) will actually use it.
However I wasn't able to find a solution which similar robustness and
cleanness.

One possible alternative implementation, which I will record here for
posterity, is to store it in a field in `Runtime`, e.g.
`Runtime::newTarget` before every call. If the called function needs it,
it is required to copy it to a local register first thing, before it
makes any other calls (which is not hard to do). Additionally, to avoid
memory leaks, it needs to be cleared when returning - otherwise it could
keep the callee alive potentially forever, or at least until another
call is done. I decided agains that implementation beause that messing
around with global state in various places in the VM is fragile and not
desirable at this stage.

== Interpreter Stack Frame Increase

This change ended up affecting the size of the interpreter native stack
frame, which is checked by a unit test. The threshold specified in the
test was 800 bytes, while the new frame in some configurations can be
up to 840 bytes. The size of the frame changes between versions of
Clang, GCC, etc. Quite surprisingly it also depends on GCKIND: choosing
GENERATIONAL and NONCONTIG_GENERATIONAL results in different sizes.

In a local Mac build the increase from this diff was exactly 48 bytes.
I verified that it is caused solely by the extra parameter to
`StackFramePtr::initFrame()` and not by adding locals in
Interpreter.cpp. Inspection of the generated assembly did not provide an
obvious explanation or a way to reverse it.

Since the chosen threshold of 800 bytes was arbitrary, we have to live
with this increase for now, but we have to be cognizant that small
changes can have unexpected impact on native stack frame size.

Reviewed By: avp

Differential Revision: D13595848

fbshipit-source-id: e499adf907c571f8584b0016ae1ba6094a70d49b