define how to implement instructions out-of-line
Summary:
Many instructions do not need to be implemented directly inside the
interpreter loop - they are either called too infrequently, or are so
slow and complex that it doesn't matter.

Define a macro `CASE_OUTOFLINE()` to automatically handle such
instructions by automatically calling an apropriately named function
with a fixed signature.

Re-implement directEval(), which was already out of line, with the new
technique.

Reviewed By: ridiculousfish

Differential Revision: D13656800

fbshipit-source-id: 64ac123330ab37063813734c15f14c46a6947a50