use getChangedOperands() in spill registers
Summary:
I noticed that SpillRegisters uses an explicit enumeration to check
which instructions modify their operands. We already have an interface
that does that: `Instruction::getChangedOperands()`, so there is always
a risk that when updating one, the other might become stale.

Change SpillRegisters to use `getChangedOperands()`. This will have some
performance cost, but I believe the additional robustness is worth it.

Updated `getChangedOperands()` to return a bitset instead of a list of
values, since the former is cheaper and functionally equivalent.

Reviewed By: dulinriley

Differential Revision: D13656798

fbshipit-source-id: ade6e1fcae6aa98e08a727cf80e0999a4d555396