New bytecode instructions for storing own properties
Summary:
This addresses two ES6-related gaps in functionality in our bytecode:
- Storing own properties by value instead of id (needed by computed
  properties in ES6)
- Creating non-enumerable own properties (needed by ES6 classes)

Add several new bytecode instructions:
- PutOwnByVal
- PutNewOwnById/Short/Long
- PutNewOwnNEById/Long
- PutOwnGetterSetterByVal

PutNewOwnById replaces the PutOwnById family, but clarifies the
semantics:
- The target must be an object.
- The property name must not be an array object.

PutNewOwnNEById is similar but creates non-enumerable properties.

PutOwnGetterSetterByVal replaces PutGetterSetter. There is no "ById"
instruction since accessors should be rare.

Extend StoreOwnPropertyInst with a flag for non-enumerability.
Introduce a new subclass of the latter, StoreNewOwnPropertyInst to
support the new assumptions of PutNewOwnById

In this commit we do not yet emit any of the new functionality, but the
ES5 code generation had to be updated to use the new instructions.

Reviewed By: avp

Differential Revision: D13669475

fbshipit-source-id: d3974218912666c014f8cdda8cd5bf49360e7ea3