support for __proto__ initializer
Summary:
The handling of __proto__ in object initializers is defined by ES6
B.3.1. It is basically required to be used exactly once as a literal
property name (not computed) and to attempt to set the prototype of the
new object, failing silently on any error.

In the VM we are adding a new instruction and a new HermesInternal call:
- NewObjectWithParent creates an object with the specified parent
- HermesInternal.silentSetPrototypeOf() sets an object parent without
  throwing exceptions on error.

In the compiler we are changing `AllocObjectInst` to take an optional
`parent` parameter.

If the object initializer sets the `__proto__` property, we want to just
create the object with the correct parent, but that may not always be
possible, in which case we use `HermesInternal.silentSetPrototypeOf()`.

Reviewed By: avp

Differential Revision: D13759358

fbshipit-source-id: ec1545f09f77bc5228af6dc6551a30557f33e1ad