introduce JSFunction::kHasFinalizer
Summary:
The knowledge whether an object has a finalizer or not is currently too
spread out and it is very easy to make a mistake, especially if rebasing
across a change that removed or added a finalizer.

In this commit we establish a new convention where the class has a
constant indicating whether it has a finalizer, and allocation functions
should just use it. The constructor enforces that the constant is
correct.

Reviewed By: avp

Differential Revision: D13663997

fbshipit-source-id: be7677c30af6baef40c09aac96d487b7b64a62b3