Add verify IR command line flag to the compiler
Summary:
Doing IR verification can take a long time if the JS file is rather large.
In order to keep dbg builds fast, we would like the option to *not* verify the IR for well-known bundles (such as the fb4a bundle).

This diff adds a command-line flag to opt-out of verification. Opting-in has no effect (and is hidden) in non-slow-debug builds.
This is mostly meant as an option to turn it off.
This is the same mechanism that handle-sanitization uses.

Reviewed By: tmikov

Differential Revision: D13910212

fbshipit-source-id: 6845ee3913b9602aa23cb83c4edaad7c84e45ab1