Introduce SafeUInt32.
Summary:
Use a SafeUInt32 which can handle overflow checking
when incrementing a size repeatedly.

Use it as a `length` input to `createStringBuilder`.
For now, this is only used in Function.cpp, but this can be reused
in any of the other places in the JS library which can overflow
the StringBuilder length.

Future plans:
- Use `SafeUInt32` everywhere sizes are incremented.
- Make `StringBuilder::createStringBuilder` only callable by using `SafeUInt32`.

Reviewed By: tmikov

Differential Revision: D13863656

fbshipit-source-id: 8a7c901598d66980a397ba57c0bf069957f18b4e