Put Runtime inside one of the storages
Summary:
The `Runtime *` needs to point at the bottom of the pre-allocated storage.
This will be used by future diffs that will make all `GCPointer` 32 bits, and add the `Runtime *` as a base of the GC.

This could be done only for `NCGen` (the other GCs don't care about this requirement); however, it doesn't matter to the other GCs where the `Runtime` is.

Reviewed By: davedets

Differential Revision: D13238190

fbshipit-source-id: 52efe21641f8187f20f84275e9c1b06b0a178441