Remove unused support for clobbering instructions
Summary:
The Hermes register allocator has a member function mayClobberRegisters(),
which allows an instruction to declare that it overwrite registers not
named as a destination. This was used by the JSC backend, but HBC has no
clobbering instructions so it always returns false.

Remove the "mayClobberRegisters()" function and simplify the one call site.

Reviewed By: willholen

Differential Revision: D13576816

fbshipit-source-id: fe452491e6385a39777b870f15009b76eb6591fa