GCSegmentRange
Summary:
Introducing a new API to represent sequences (or ranges) of segments, to replace (eventually -- not in this diff) `OldGenSegmentIterator`.

The incumbent API was not ideal because it was based on C++ iterators, which are not good at expressing early termination.  Termination is checked by comparison against a sentinel `end` value, rather than being an intrinsic property of the thing being iterated over.  This made it difficult to express the fact that we might fail to retrieve a segment from the OldGen if it had not been materialised and materialisation failed.

This new API solves that problem by having termination indicated by a sentinel value (a nullptr) being returned when we request the next segment.  I took some inspiration from kodafb's `ConsumableRange` class, and a `GCSegmentRange` can be constructed from the former (backed by an appropriate iterator), but I chose to describe it as a distinct concept because `ConsumableRange` has separate `hasNext()` and `next()` API calls.  In order to implement a class with the `ConsumableRange` interface that gets segments by materialising them on-demand, we would have to lazily materialise the segment for either the `hasNext()` or `next()` calls.  The facet of this that I did not like in particular is that a call to `hasNext()` may potentially do a lot of hard work (materialising a segment) that is wasted if we don't end up calling `next()`.  Having an API that makes no distinction between asking for the next segment and asking whether it exists makes the potential of this wasted work more obvious.

This diff only introduces the wrapper to create a segment range from a `ConsumableRange`.  It has been introduced within a namespace because other combinators for constructing segment ranges will follow.

Reviewed By: davedets

Differential Revision: D13880606

fbshipit-source-id: d3ba4c3a1a3239213cab6cab4cc36680e478ea5d