GCSegmentRange::Fused
Summary:
An adapter class that prevents a range from being able to "start-up" again once it has failed.

This is useful because a range that materialises segments could conceivably fail when queried once, but then succeed on subsequent requests.  However, when we are requesting segments to compact into during full collection, if we fail to materialise a segment in the old generation, we want to move on and never try to materialise a segment again, not only because it's potentially expensive to do that, but also because if we do succeed in materialising a segment after the fact, we will be compacting objects out of order.

Reviewed By: davedets

Differential Revision: D13881716

fbshipit-source-id: eb4c2f8c7799c155860a33e42dc75317fdef5a8b