GCSegmentRange::Concat
Summary:
Combinator for concatenating segment ranges together.  The implementation maintains a list of delegate ranges and a cursor over the currently active one.  The cursor moves when the delegate under it fails.

I also trialled an implementation using templates such that the body of `next()` for the concatenation of three ranges `R0`, `R1`, `R2` was equivalent to the following:

```
AlignedHeapSegment *next() {
  if (auto *res = R0.next()) {
    return res;
  }

  if (auto *res = R1.next()) {
    return res;
  }

  if (auto *res = R2.next()) {
    return res;
  }

  return nullptr;
}
```

Similar to the approach that LLVM's `concat_iterator` takes.  Call the solution presented in this diff `RangeVirtual` and the alternative `RangeStatic`, then in tests on my local machine:

 - `RangeStatic` takes 0.373ns to construct, 252.135ns to build and iterate through completely.  Source: P60881283
 - `RangeVirtual` takes 559.349ns to construct, 1172.890ns to build and iterate through completely.  Source: P60881284

In all cases, I'm taking the minimum of three runs, and in each run I performed 10 warm-up iterations followed by 1,000,000 repetitions, taking the average across the latter reps.  Run-to-run variance was low.  Results were corroborated by runs on an ARMv7 Android device as well, so this is not architecture specific.

The way concatenation will be used (eventually, not in this diff) will result in concatenations being components of other concatenations.  `RangeStatic` has a flattening behaviour:  The compiler emits equivalent assembly for `Concat<Concat<X, Y>, Z>` and `Concat<X, Y, Z>` as evidenced by https://godbolt.org/z/TFEBiw (The example in the Godbolt actually has 4 components, to reflect the structure we will actually see when using Concat on segment ranges).

I did this for my own curiosity -- wanting to get a good idea what the cost of these abstractions looked like in practise -- rather than because the code needed this level of performance tuning.  Despite being better in the microbenchmark, when integrated into the GC (as in this diff), performance was a wash, so I have implemented the simpler version, using virtual dispatch.

Reviewed By: davedets

Differential Revision: D13880605

fbshipit-source-id: 7db1a5ab2781d129f1d0a0fdc162371ae7568f8a