Fix "require" failure on Windows
Summary:
require uses slash regardless of platform.
llvm::sys::path functions by default chooses the kind of slash automatically
based on platform. As a result, the helper functions changes slash direction
in module paths automatically, which leads to failures to lookup modules on
Windows.

Ideally, module path (in the abstracted sense) and file path (in the physical
sense) should never be used interchangeably, and be treated as if they are
different types. However, this is unnecessarily because Windows also accepts
slash in file paths.

Reviewed By: tmikov

Differential Revision: D14014672

fbshipit-source-id: 31dfa515c1e78353c14c6c00de36e2bbb442607c