Another try at moving LLVM into hermes/facebook.
Summary:
This is a modified version of D13600791.

We copy LLVM into hermes/facebook, and make all the references in hermes refer to that version of LLVM.

In contrast to D13600791:
 * We leave hcjsc-deprecated unmodified.  It continues to reference xplat/third-party/llvm.  Marc noted in D13600791 that this is used only by an Oculus AR thing, which is switching to hermes in a relatively short time.  So leave it as is, and let that reference to it go away, and delete it, decreasing the "ref count" on third-party/llvm by one.
 * We leave hermes-snapshot unmodified.  It continues to referenc xplat/third-party.  A follow-on diff to this will take a snapshot, which presumably will overwrite the hermes-snapshot BUCK files with those from hermes.  These will reference LLVM as a hermes_dep to facebook/llvm.  Presumably this will cause them to reference hermes-snapshot/facebook/llvm.  So that snapshot will copy llvm iinto that location in hermes-snapshot.
* We modify the specifications of the native .so's to have a two copies of LLVM.  These will almost always be the same, so it won't cost any compressed space.

This puts us in a place where we can update LLVM in the hermes part, but leave the hermes-snapshot LLVM unmodified.
Without doing this, upgrading LLVM would require modifying the hermes-snapshot code...which would make it not so much of a snapshot.

Ideally, we would just include the LLVM's as part of libhermes(-snapshot).so.  When I tried this, however, I got signficantly larger APK size.  For reasons I don't yet understand, including LLVM in libhermes.so disables some "pruning" that had otherwise happened; the size of the LLVM files is much larger than when in libhermes-thirdpartydeps.so, because many more symbols are included.  If we can figure that out, and get the same set of symbols included, that would be a cleaner solution.

Reviewed By: mhorowitz

Differential Revision: D13966934

fbshipit-source-id: c4b4b5041f1c1b3fa54d5996f8a5e622678aaa82