Handle non-ASCII character in argv
Summary:
When main is invoked on Windows, non-ASCII characters is automatically converted
to iso-8859-1 (or maybe some other ASCII-like encoding) on a Windows machine
with US locale. This conversion is lossy and irreversible.

LLVM provides a cross-platform helper to retrieve argv that is UTF-8 encoded.
On Windows, it manages to do this by making a system call to retrieve the argv,
ignoring the ones passed into main.

Reviewed By: ridiculousfish

Differential Revision: D14075290

fbshipit-source-id: 785571565a132a98ee70997525dd328ae1fea6ea