Fix leak of JSError stack traces.
Summary:
In fbcode, the Hermes Service tests, when run in dev mode, which does leak detection, reported leaks.  These were all of stack traces recorded in JSError objects.  I tracked this down to use of release instead of reset on std::unique_ptr.  Fix is simple.

D14071993 fixes this in fbcode.  This diff copies that fix to xplat.

Differential Revision: D14072090

fbshipit-source-id: ba28ea9fc3e8b3e2fac861ec5598d0da6ba3610b