The key is to ensure there is no key
Summary:
Apologies for being so vague, but vagueness is what is required now.

Mark BCProviderBase's one out-of-line virtual function as non-virtual (it
was needlessly virtual). This ensures the class has no key function so
linkage for its vtable becomes vague. This means that API/hermes.cpp (which
uses rtti) will emit typeinfo for it and we will link once again. This fixes
the CMake build.

Reviewed By: tmikov

Differential Revision: D14116607

fbshipit-source-id: ac91c1ea39fbc72d07202c793328d8c8c03438e1