fix StackRuntime rtti link error
Summary:
StackRuntime didn't have any virtual methods, so some compilers would attempt to
generate rtti info for it in a compilation unit that uses it, if that unit was
compiled with -frtti. To avoid that, add an empty virtual destructor and
provide a body in Runtime.cpp, which is compiled without rtti.

Reviewed By: davedets

Differential Revision: D14161033

fbshipit-source-id: 6c463002b242679274b907b982b85df423392c3e