Fix MallocGC to recognize that shouldSanitizeHandles is not deterministic.
Summary:
Originally, shouldSanitizeHandles returned a boolean that was fixed in the GCConfig: we were either always sanitizing handles, or never doing so.  At some point in the recent past, we changed this to do "sampling".  The intention was that we would specify a "sanitize handle probability" -- instead of doing a relocating GC at each allocation when sanitizing was enabled, we would do so with probability p, specified on the command line.

The problem (identified by RIley) was that the MallocGC code was gating things that should *always* be done whenever handle sanitization was *possible* where instead calling "shouldSanitizeHandles" to determine whether they should be done!  (I discovered this by finding that running with -gc-sanitize-handles=1.0 worked, but -gc-sanitize-handles=0.9999 didn't, for some example!)

The fix is to do those things that should always be done whenever handle sanitization is possible under the build flag.  This means that we'll always be doing that extra, even when -gc-sanitize-handles=0.0.  We could add a dynamic test for whether the p is > 0.0 -- but must of the time when handle sanitization is enabled in the build, we want to enable it dynamically.

Reviewed By: kodafb

Differential Revision: D14174884

fbshipit-source-id: 135e484ab238e7c162af1e512ee35e57add4b08f