Make SynthTrace use decorator pattern
Summary:
The main goals here are to move the synth stuff out of hermes/hermes.h
so :HermesAPI doesn't depend on any of the trace stuff at all, and to
minimize synth's deps on HermesRuntime.

Since HERMSESMV_API_TRACE is a compile-time option not exposed outside
of this directory, the way we do this is by introducing a factory
`makeTracingHermesRuntime` in a new header `hermes/hermes_tracing.h`.
This will return a HermesRuntime decorated to enable tracing if trace
is enabled, and just the plain HermesRuntime if not.  Note that the
return type is jsi::Runtime, which erases the tracing-ness.

notable impl:

most code moved inside facebook::hermes::tracing namespace
TracingRuntime decorator split out of hermes.h

Reviewed By: dulinriley

Differential Revision: D13962150

fbshipit-source-id: 864a8e39394e7a0ba727567434f7e35a76d9df9f