Copy updated LLVM into hermes/facebook/llvm (copy, llvm mods, + hermes mods)
Summary:
We have decided to solve the problem of compatiblity of the xplat and fbcode versions of Hermes by committing to track the LLVM version used in fbcode in xplat.

This diff is the "hg fold" of the revisions for the three diffs:
D14135592, D14136068, and D14111772.

This diff updates the LLVM version to revision
4519ac3791135eb9c207f0684f4236dbc13ac83f
in the LLVM git repo  This is the version currently in fbcode/third-party2/llvm-fb/stable.

In more detail:
* The first diff in the stack just copies the new LLVM into hermes/facebook/lllvm.
* The second diff makes a small set of changes to the LLVM source.
* The third diff embodies the small set of source changes in Hermes necessary to use the new LLVM APIs.  One of these, in Handle.h, is a workaround for an LLVM shortcoming -- it's marked with a TODO, with a task, to be reconsidered if LLVM changes this issue next version.

Note that the three diffs are accepted; this diff should not require further review.

Note: we expect this to have an app size regression, because the LLVM versions used by hermes and hermes-snapshot will no longer be identical.  However, this situation should be temporary; when we next take a snapshot, they will again be the same, and the app size regression should go away.

allow-large-files

Reviewed By: dulinriley

Differential Revision: D14236782

fbshipit-source-id: 7a196ff25b258e8a05993cd5c4dff79d27f9a55e