Fix bug in TraceInterpreter, always set the mocked environment
Summary:
In a stack of diffs made to convert the tracing system to use a decorator instead of inheritance, the functions to create a hermes runtime were changed.

In one path (when tracing is on), the mocked environment is being set.
In the other path, the mocked environment was *not* being set.
This resulted in non-determinism from `Date.now()` and `Math.random()`, which caused some strange bugs to occur.

In a follow-up diff, I'll make this impossible to occur and add a verification exception to the TraceInterpreter.

Reviewed By: kodafb

Differential Revision: D14264327

fbshipit-source-id: 0d583c5a63ded05eb57ae8302cacac6925c47998